Pull ASN treemap into a separate component
